feat: add validation for blog update delete and user block

+++ src/app/modules/admin/admin.service.ts
@@ -1,7 +1,13 @@
+import { httpStatus } from "../../config/httpStatus";
+import AppError from "../../errors/AppError";
 import { Blog } from "../blog/blog.model";
 import { User } from "../user/user.model";
 
 const blockUserIntoDB = async (id: string) => {
+    const user = await User.isUserExistsById(id);
+    if (!user) {
+        throw new AppError(httpStatus.NOT_FOUND, "Requested user is not found");
+    }
     const result = await User.findByIdAndUpdate(
         id,
         { isBlocked: true },
@@ -12,6 +18,10 @@ const blockUserIntoDB = async (id: string) => {
 };
 
 const deleteABlogFromDB = async (id: string) => {
+    const blog = await Blog.isBlogExists(id);
+    if (!blog) {
+        throw new AppError(httpStatus.NOT_FOUND, "Requested blog is not found");
+    }
     const result = await Blog.findByIdAndDelete(id);
     return result;
 };
